refactor(structure): move games and project components into a components folder for better clarity and structure
