+++ src/App.tsx
@@ -2,8 +2,8 @@ import { BrowserRouter } from "react-router-dom";
 import "./App.css";
 import AppRoutes from "./routes";
 import { GameProgressProvider } from "./context/GameprogressProvider";
-import Progressbar from "./assets/Progressbar/Progressbar";
-import Header from "./assets/Header/Header";
+import Progressbar from "./components/Progressbar/Progressbar";
+import Header from "./components/Header/Header";
 
 function App() {
     return (
