+++ src/components/Game1/Game1.css
