+++ src/components/Game2/Game2.tsx
