+++ src/routes.tsx
@@ -1,10 +1,10 @@
 import { Route, Routes } from "react-router-dom";
-import Game1 from "./assets/Game1/Game1";
-import Game2 from "./assets/Game2/Game2";
-import Game3 from "./assets/Game3/Game3";
-import Game4 from "./assets/Game4/Game4";
-import Game5 from "./assets/Game5/Game5";
-import Gameselection from "./assets/Gameselection/Gameselection";
+import Game1 from "./components/Game1/Game1";
+import Game2 from "./components/Game2/Game2";
+import Game3 from "./components/Game3/Game3";
+import Game4 from "./components/Game4/Game4";
+import Game5 from "./components/Game5/Game5";
+import Gameselection from "./components/Gameselection/Gameselection";
 
 function AppRoutes() {
     return (
